Rename: renamed 'config-messages' to 'settings-menus'
